requirements.txt
snowflake-connector-python>=4.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
google-generativeai>=0.8.0
python-multipart>=0.0.12
websockets>=13.0
httpx>=0.28.0
python-dotenv>=1.0.0
plotly>=5.24.0
sphinx-ai-cli>=1.0.6
